Add files via upload
Setting for data generation
1. # clusters = 50, 50 people/cluster (i.e. let’s focus on this setting to make sure that the program)
2. Covariate X generation: make sure that the variance of X is not too large i.e. at the value Fan used in his code
3. Outcome model S1 i.e. no interaction and a single ICC for this i.e. 0.05
4. Missing model
a. Miss 2 i.e. no clustering 
b. Miss 5 with different ICC of missingness: 0.1,0.2,0.3,0.4,0.5,0.6,0.8,0.9 but with the overall proportion missing set to be approx. 30% i.e. might need to change the intercept in the logistic model
5. Calculate the truth using Hosain’s method i.e. analyzing the complete data set (not Melanie’s approach from the paper)

Analysis of each data set
1. Only first do IPW (not MMI)
2. For IPW, use only Melanie’s package not geepack.
3. Run W-GEE using both independence and exchangeable in each scenario (up to now, we haven’t seen there be much difference between these two) with the goal that we do not need to use stabilization to remove extreme weights i.e. we want to make sure tha data generation is such that we do not have extreme weights. If we do, we will change the data generation models.
4. IPW versions
a. Weights calculated with no clustering and correct model i.e. A+X
b. Weights calculated with clustering and correct model i.e. A+X + theta